New package by Divergent Insights
We are excited of making available to the community our very first dbt package. This package helps to create, configure and destroy Snowflake environments. This initial version provides a basic setup from which many other will be derived (Snowflake environment setup can be very opinionated so flexibility is important here)